range charts changes
