+++ lib/index.js
@@ -158,9 +158,12 @@ export const Dataviz = function (options = {}) {
     this.config.range = true;
     this.config.type = this.config.type.replace('-range', '');
     this.config.legend.show = false;
-    this.config.labels = ['Max', 'Min'];
     this.config.stacking = 'normal';
-    this.config.colors = ['#ff0000', '#ff0000'];
+    this.config.labels = ['Max', 'Min'];
+    this.config.colors = [
+      this.config.colors[0],
+      this.config.colors[0],
+    ];
   }
 
   // overwriting stacked when stacking added
@@ -691,13 +694,6 @@ Dataviz.prototype.render = function(results = undefined){
       }
     }
   });
-  // range charts cheat for Min stacked values visibility
-  if(datavizInstance.config.range){
-    if(datavizInstance.config.type.includes('bar')){
-      return document.querySelector('.c3-target-Min').style.display = 'none';
-    }
-    return document.querySelector('.c3-area-Min').style.display = 'none';
-  }
 
   return returnValue;
 };
@@ -817,11 +813,14 @@ Dataviz.version = pkg.version;
 function buildDomWrapper(el, options) {
   let html = '';
   let chart100percentWide = '';
+  let rangeChart;
   if (options.legend.position === 'top' || options.legend.position === 'bottom') {
     chart100percentWide = 'c3-chart-100-percent';
   }
-
-  let container = `<div class="c3-chart ${chart100percentWide}"></div>`;
+  if (options.range) {
+    rangeChart = 'keen-dataviz-range';
+  }
+  let container = `<div class="c3-chart ${chart100percentWide} ${rangeChart}"></div>`;
   let align = 'horizontal';
   if (options.legend.position === 'left' || options.legend.position === 'right') {
     align = 'vertical';
