+++ lib/libraries/index.js
@@ -8,6 +8,9 @@ import c3DefaultDateFormat from './c3/extensions/default-date-format';
 import c3PaginatingLegend from './c3/extensions/paginating-legend';
 import c3TooltipContents from './c3/extensions/tooltip-contents';
 
+import calculateRange from '../dataset/utils/calculate-range';
+import { calculateSumForPercents, calculatePercents } from '../dataset/utils/calculate-percents';
+
 import message from './default/message';
 import metric from './default/metric';
 import table from './default/table';
@@ -479,43 +482,4 @@ function bindResizeListener(fn){
   else if (window.attachEvent) {
     window.attachEvent('onresize', fn);
   }
-}
-
-function calculateSumForPercents(matrix) {
-  const sumArray = [];
-  matrix.slice(1).forEach((d, i) => {
-    d.forEach((e) => {
-      if (typeof e === 'number') {
-        if (!sumArray[i]) {
-          sumArray[i] = e;
-          return sumArray[i];
-        }
-        sumArray[i] += e;
-      }
-      return sumArray[i];
-    });
-  });
-  return sumArray;
-}
-
-function calculatePercents(matrix, sumArray) {
-  const newValues = matrix.slice(1).map((d, i) => d.map((e) => {
-    if (typeof e === 'number') {
-      return (e / sumArray[i]) * 100;
-    }
-    return e;
-  }));
-  return newValues;
-}
-
-function calculateRange(matrix) {
-  const newValues = matrix.slice(1).map(d => d.map((e, i) => {
-    if (typeof e === 'number') {
-      if (i === 1) {
-        return e - d[i + 1];
-      }
-    }
-    return e;
-  }));
-  return newValues;
 }
